feat: Add build and test workflows for PogRust
This commit adds two new workflow files for building and testing PogRust development image (`dev-pogrust-build.yml`) and publishing PogRust image (`main-pogrust-build-publish.yml`). The workflows are triggered on a schedule and for specific branches and paths. The Docker images are built using Docker buildx and pushed to a Docker registry. The workflows also include steps for logging into the registry, extracting metadata, and signing the published Docker image. Additionally, a Dockerfile for PogRust is included, which sets up a Rust development environment and adds a user for development purposes.